refactor : clean file structures with architecture
