+++ src/app/layouts/index.js
@@ -1,0 +1,2 @@
+export { default as AppFrame } from './AppFrame';
+export { default as RestaurantDetailFrame } from './RestaurantDetailFrame';
