+++ src/features/review/components/index.js
@@ -1,0 +1,4 @@
+export { default as ReviewInput } from './ReviewInput';
+export { default as ReviewItem } from './ReviewItem';
+export { default as ReviewPagination } from './ReviewPagination';
+export { default as ReviewStarRatingSelector } from './ReviewStarRatingSelector';
