+++ src/features/review/constants/PaginationSettings.jsx
@@ -1,4 +1,4 @@
 export const PAGE_OPTIONS = {
   ITEMS_PER_PAGE: 5,
   PAGES_PER_GROUP: 5,
-}
+};
