+++ src/pages/index.js
@@ -1,0 +1,3 @@
+export { default as HomePage } from './HomePage';
+export { default as ErrorPage } from './ErrorPage';
+export { default as RestaurantDetailPage } from './RestaurantDetailPage';
