+++ src/shared/apis/axios.js
@@ -1,10 +1,8 @@
 import axios from 'axios';
 
-const instance = axios.create({
+export const instance = axios.create({
   baseURL: 'https://kiryong.site/',
   params: {
     language: 'ko-KR',
   },
 });
-
-export default instance;
