+++ src/shared/apis/index.js
@@ -1,0 +1,2 @@
+export { instance as axios } from './axios';
+export { requests } from './requests';
