+++ src/shared/assets/index.js
@@ -1,0 +1,2 @@
+export { default as NoticeImg } from './kiryong-notice.svg';
+export { default as RiceImg } from './rice.svg';
