+++ src/shared/datas/index.js
@@ -1,0 +1,3 @@
+export { default as menus } from './menus.json';
+export { default as restaurantInfo } from './restaurantInfo.json';
+export { default as reviews } from './reviews.json';
