+++ src/shared/index.js
@@ -1,0 +1,5 @@
+export * from './apis';
+export * from './assets';
+export * from './datas';
+export * from './stores';
+export * from './utils';
