+++ src/shared/stores/index.js
@@ -1,0 +1,1 @@
+export { useSelectedRestaurantStore } from './useSelectedRestaurantStore';
