+++ src/shared/utils/index.js
@@ -1,0 +1,3 @@
+export { fetchData } from './fetchData';
+export { getTodayDate } from './getTodayDate';
+export { setDormMenuData } from './setDormMenuData';
