+++ src/widgets/Header.jsx
@@ -1,79 +1,26 @@
-import styled from '@emotion/styled'
-import { useLocation } from 'react-router-dom'
-import { NavLink } from 'react-router-dom'
+import styled from '@emotion/styled';
+import { useLocation, NavLink } from 'react-router-dom';
 
-function Header() {
-  const location = useLocation()
+export default function Header() {
+  const location = useLocation();
 
   const getBgColor = () => {
     switch (location.pathname) {
       case '/':
-        return '#f4f4f4'
+        return '#f4f4f4';
       case '/restaurant':
-        return '#333'
+        return '#333';
     }
-  }
+  };
 
   const getTxtColor = () => {
-    if (location.pathname === '/restaurant') return '#fff'
-    else return '#000'
-  }
+    if (location.pathname === '/restaurant') return '#fff';
+    else return '#000';
+  };
 
-  const Header = styled.header`
-    background-color: ${(props) => props.bgcolor};
-    color: ${(props) => props.txtcolor};
-    height: 100px;
-    padding: 0px 40px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    @media (max-width: 480px) {
-      height: 70px;
-    }
-  `
-  const Logo = styled(NavLink)`
-    color: ${(props) => props.txtcolor};
-    font-size: 30px;
-    font-weight: bold;
-    text-decoration: none;
-    @media (max-width: 480px) {
-      font-size: 25px;
-    }
-  `
-  const SmallLogo = styled.span`
-    margin-left: 10px;
-    @media (max-width: 480px) {
-      display: none;
-    }
-  `
-  const MenuBar = styled.div`
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    margin: 0;
-  `
-  const StyledNavLink = styled(NavLink)`
-    color: ${(props) => props.txtcolor};
-
-    text-decoration: none;
-    margin-left: 30px;
-
-    &.active {
-      color: #00abaa;
-    }
-
-    &:hover {
-      color: #00abaa;
-    }
-
-    @media (max-width: 480px) {
-      margin-left: 10px;
-      font-size: 12px;
-    }
-  `
   return (
     <>
-      <Header bgcolor={getBgColor()} txtcolor={getTxtColor()}>
+      <HeaderDiv bgcolor={getBgColor()} txtcolor={getTxtColor()}>
         <div>
           <Logo txtcolor={getTxtColor()} to="/">
             기밥
@@ -88,9 +35,60 @@ function Header() {
             개발자들
           </StyledNavLink>
         </MenuBar>
-      </Header>
+      </HeaderDiv>
     </>
-  )
+  );
 }
 
-export default Header
+const HeaderDiv = styled.header`
+  background-color: ${(props) => props.bgcolor};
+  color: ${(props) => props.txtcolor};
+  height: 100px;
+  padding: 0px 40px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  @media (max-width: 480px) {
+    height: 70px;
+  }
+`;
+const Logo = styled(NavLink)`
+  color: ${(props) => props.txtcolor};
+  font-size: 30px;
+  font-weight: bold;
+  text-decoration: none;
+  @media (max-width: 480px) {
+    font-size: 25px;
+  }
+`;
+const SmallLogo = styled.span`
+  margin-left: 10px;
+  @media (max-width: 480px) {
+    display: none;
+  }
+`;
+const MenuBar = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  margin: 0;
+`;
+const StyledNavLink = styled(NavLink)`
+  color: ${(props) => props.txtcolor};
+
+  text-decoration: none;
+  margin-left: 30px;
+
+  &.active {
+    color: #00abaa;
+  }
+
+  &:hover {
+    color: #00abaa;
+  }
+
+  @media (max-width: 480px) {
+    margin-left: 10px;
+    font-size: 12px;
+  }
+`;
