+++ src/widgets/InnerContentCard.jsx
@@ -1,19 +1,5 @@
 import styled from '@emotion/styled';
 
-export const InnerContentWrapper = styled.div`
-  width: 100vw;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: row;
-  justify-content: center;
-  margin: 10px 0;
-
-  @media (max-width: 480px) {
-    flex-direction: column;
-  }
-`;
-
 export const InnerContentCard = styled.div`
   display: flex;
   box-sizing: border-box;
