+++ src/widgets/InnerContentWrapper.jsx
@@ -1,0 +1,15 @@
+import styled from '@emotion/styled';
+
+export const InnerContentWrapper = styled.div`
+  width: 100vw;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: row;
+  justify-content: center;
+  margin: 10px 0;
+
+  @media (max-width: 480px) {
+    flex-direction: column;
+  }
+`;
