+++ src/widgets/index.js
@@ -1,0 +1,6 @@
+export { default as Header } from './header';
+export { InnerContentWrapper } from './InnerContentWrapper';
+export { InnerContentCard } from './InnerContentCard';
+export { default as InnerTitlesWrapper } from './InnerTitlesWrapper';
+export { Selector } from './Selector';
+export { SelectorButton } from './SelectorButton';
